src/requirements.txt
Django==3.0.4
djangorestframework==3.11.0
markdown==3.2.1
django-filter
django-environ
gunicorn==19.9.0